Smart fix for Adorner updating on size changes
What it says on the tin!
TODO: Probably best to create 'AdornerContainer' class to replace the Border and hold the AdornedElement with a WeakRef, use WeakRefListener for SizeChanged of the underlying adorned element
Think SizeChanged reset is the best approach, as binding I don't think gets us much as helper in toolkit is just listening to SizeChanged anyway.
Could think of CanvasView as well, but think this works fine for simple approach for now.
Should move to Toolkit and work on AdornerContainer there, but this POC seems to work great.